# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: "Vending machine Swagger Docs"
  description: "Everything related to vending machine test"
servers:
  - description: Local development
    url: "http://localhost:3000"
paths:
...
  /auth/register:
    post:
      tags:
        - registration
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        201:
          $ref: "#/components/responses/RegisterUser"
        400:
          $ref: "#/components/responses/Error400"
        401:
          $ref: "#/components/responses/Error401"
        404:
          $ref: "#/components/responses/Error404"
        409:
          $ref: "#/components/responses/Error409"
        500:
          $ref: "#/components/responses/Error500"
  /auth/login:
    post:
      tags:
        - login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        201:
          $ref: "#/components/responses/Login"
        400:
          $ref: "#/components/responses/Error400"
        401:
          $ref: "#/components/responses/Error401"
        404:
          $ref: "#/components/responses/Error404"
        500:
          $ref: "#/components/responses/Error500"
...
components:
  schemas:
...
    RegisterUserRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
        - role
      properties:
        username:
          type: string
          example: "JohnSnow"
        password:
          type: string
          format: password
          minLength: 8
          example: "abcd1234"
        role:
          type: string
          items:
            type: string
            enum: ["BUYER", "SELLER"]
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "userName1"
        password:
          type: string
          format: password
          minLength: 8
          example: "abcd1234"
...
    UserRole:
      type: string
      enum: ["BUYER", "SELLER"]
    Error400Response:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - API_VALIDATION_ERROR
            - INVALID_JSON_FORMAT
        message:
          type: string
        errors:
          $ref: "#/components/schemas/ArrayOfErrors"
    Error401Response:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - UNAUTHORIZED_REQUEST
        message:
          type: string
...
    Error404Response:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - USER_NOT_FOUND
            - PRODUCT_NOT_FOUND
        message:
          type: string
    Error409Response:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - DUPLICATE_END_USER
        message:
          type: string
...
    Error500Response:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
            - INTERNAL_UPSTREAM_ERROR
        message:
          type: string
    ArrayOfErrors:
      nullable: true
      type: array
      items:
        oneOf:
          - type: string
          - type: object
  responses:
    RegisterUser:
      description: User registration response
      content:
        application/json:
          schema:
            type: object
            required:
              - id
              - username
              - role
              - token
            properties:
              id:
                type: string
                example: "eu-cc050ed9-a39f-4bd8-a4aa-2e72c9638d3f"
              username:
                type: string
                example: "username"
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdGF0dXMiOiJWRVJJRklFRCIsImVuZF91c2VyX2lkIjoiZXUtMSJ9._AJoQ7FVsVrQByrNs9oKwsZ67uMMkHa4qZYuihDhfTo
              role:
                $ref: "#/components/schemas/UserRole"
    Login:
      description: Login response
      content:
        application/json:
          schema:
            type: object
            required:
              - id
              - username
              - role
              - token
            properties:
              id:
                type: string
                example: "eu-cc050ed9-a39f-4bd8-a4aa-2e72c9638d3f"
              username:
                type: string
                example: "username"
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdGF0dXMiOiJWRVJJRklFRCIsImVuZF91c2VyX2lkIjoiZXUtMSJ9._AJoQ7FVsVrQByrNs9oKwsZ67uMMkHa4qZYuihDhfTo
              role:
                $ref: "#/components/schemas/UserRole"
    Error400:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error400Response"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error401Response"
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error404Response"
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error409Response"
...
    Error500:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error500Response"